pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-supportdesk-env"
version = "0.1.0"
description = "Support desk triage environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openenv-core @ git+https://github.com/meta-pytorch/OpenEnv.git@main",
  "fastapi>=0.115.0",
  "openai>=1.54.0",
  "pydantic>=2.9.0",
  "requests>=2.32.0",
  "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
server = "supportdesk_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["supportdesk_env", "supportdesk_env.server"]
package-dir = { "supportdesk_env" = ".", "supportdesk_env.server" = "server" }
